+++ packages/runtime/id-compressor/src/test/index.ts
@@ -1,0 +1,10 @@
+/*!
+ * Copyright (c) Microsoft Corporation and contributors. All rights reserved.
+ * Licensed under the MIT License.
+ */
+
+/**
+ * Exports for `id-compressor`
+ */
+
+export { createAlwaysFinalizedIdCompressor } from "./idCompressorTestUtilities";
